// vehicleMaintenance/src/app/common/models/globals.ts
export interface VehicleDetail {
    Make_ID: {label:string ,value:number}
    Make_Name: {label:string ,value:string}
    Model_ID: {label:string ,value:number}
    Model_Name: {label:string ,value:string}
    VehicleTypeId: {label:string ,value:number}
    VehicleTypeName: {label:string ,value:string}

}